gradient-service-domain/requirements/requirements.txt
attrs==21.2.0
bleach==4.1.0
boto3==1.18.26
botocore==1.21.61
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
cmarkgfm==0.6.0
colorama==0.4.4
cryptography==3.4.8
docutils==0.17.1
grpcio==1.33.2
grpcio-tools==1.33.2
idna==2.10
importlib-metadata==4.8.1
iniconfig==1.1.1
invoke==1.6.0
jeepney==0.7.1
jmespath==0.10.0
keyring==23.2.1
more-itertools==8.10.0
packaging==21.0
pip==21.3
pkginfo==1.7.1
pluggy==1.0.0
protobuf==3.18.1
py==1.10.0
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==5.4.1
readme-renderer==27.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
s3transfer==0.5.0
SecretStorage==3.3.1
setuptools==58.2.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
twine==0.0.0
urllib3==1.26.7
webencodings==0.5.1
wheel==0.37.0
zipp==3.6.0

